+++ refena/lib/src/provider/types/view_provider.dart
@@ -1,4 +1,3 @@
-import 'package:meta/meta.dart';
 import 'package:refena/src/notifier/base_notifier.dart';
 import 'package:refena/src/provider/base_provider.dart';
 import 'package:refena/src/provider/override.dart';
@@ -9,18 +8,28 @@ import 'package:refena/src/ref.dart';
 /// A common use case is to define a view model that depends on many providers.
 /// Don't worry about the [ref], you can use it freely inside any function.
 /// The [ref] will never become invalid.
+///
+/// Set [describeState] to customize the description of the state.
+/// See [BaseNotifier.describeState].
+///
+/// Set [debugLabel] to customize the debug label of the provider.
 class ViewProvider<T> extends BaseWatchableProvider<ViewProviderNotifier<T>, T>
     with ProviderSelectMixin<ViewProviderNotifier<T>, T> {
-  @internal
-  final T Function(WatchableRef ref) builder;
+  final T Function(WatchableRef ref) _builder;
+  final String Function(T state)? _describeState;
 
-  ViewProvider(this.builder, {String? debugLabel})
-      : super(debugLabel: debugLabel ?? 'ViewProvider<$T>');
+  ViewProvider(
+    this._builder, {
+    String Function(T state)? describeState,
+    String? debugLabel,
+  })  : _describeState = describeState,
+        super(debugLabel: debugLabel ?? 'ViewProvider<$T>');
 
   @override
   ViewProviderNotifier<T> createState(Ref ref) {
     return ViewProviderNotifier<T>(
-      builder,
+      _builder,
+      describeState: _describeState,
       debugLabel: customDebugLabel ?? runtimeType.toString(),
     );
   }
@@ -35,6 +44,7 @@ class ViewProvider<T> extends BaseWatchableProvider<ViewProviderNotifier<T>, T>
       provider: this,
       createState: (_) => ViewProviderNotifier(
         builder,
+        describeState: _describeState,
         debugLabel: customDebugLabel ?? runtimeType.toString(),
       ),
     );